+++ src/KaTeX/VirtualDomDefinitions.cs
